8 basic algorithms have been developed for solving linear equations.
The linear equations solving algorithms have been developed in C based
on the well-known Gauss elimination method and the matrix LU
decomposition. It is remarked that here a specified subclass of linear
systems are considered where the row and the colunm dimensions of the
coefficient matrix of a linear system is assumed to be equal.